requirements.txt
django
python-dotenv==0.20.0
numpy
Pillow
reportlab
psycopg2-binary
supabase